restructure for better overview
