+++ src/routes/mod.rs
@@ -1,0 +1,3 @@
+pub mod health_check;
+
+pub use health_check::*;
